# Use FastGaussQuadrature for transform between Position and Fock bases