This implements transform between position and fock via FastGaussQuadrature following the recommendation in the comments of #137: https://github.com/qojulia/QuantumOpticsBase.jl/pull/137#issuecomment-1719150078