@akirakyle 2023-09-14
Another reason to prefer this implementation is because FastGaussQuadrature's `hermpoly_rec` seems to also be faster. For the example in https://github.com/qojulia/QuantumOpticsBase.jl/pull/137#issue-1893184101 I get 

Before this PR I get for the first run

```
 0.213219 seconds (128.41 k allocations: 39.285 MiB, 7.49% gc time, 42.21% compilation time)
```

Then subsequent runs I get

```
0.151279 seconds (110 allocations: 30.717 MiB, 8.53% gc time)
```

After this PR I get first run

```
0.184702 seconds (293.64 k allocations: 287.847 MiB, 21.37% gc time, 47.35% compilation time)
```

Then subsequent runs I get

```
0.097279 seconds (140.01 k allocations: 277.558 MiB, 28.39% gc time)
```