@Krastanov 2023-09-17
This pull request looks great. 100% diff coverage, no change to tests, no JET warnings, a very reasonable new dependency (small, stable, mature, used by many others). Now we have less code to support and have better numerical stability and better performance.

A minor issue is that we seem to be using an internal method of that dependency, but the CI will catch if that is an issue.

Another issue: the number of invalidation at loading this library had doubled, to 400 with this PR. This can be very bad for loading time and TTFX. As Akira reported above, TTFX is actually fine. Concerning loading:

master on julia 1.9.3:
```
julia> @time @eval using QuantumOpticsBase
  0.255590 seconds (536.85 k allocations: 39.443 MiB, 20.51% gc time, 26.03% compilation time: 24% of which was recompilation)
```

this PR on julia 1.9.3:
```
julia> @time @eval using QuantumOpticsBase
  0.448067 seconds (1.43 M allocations: 100.047 MiB, 14.19% gc time, 12.30% compilation time: 29% of which was recompilation)
```

master on julia nightly (17 Sep 2023):
```
julia> @time @eval using QuantumOpticsBase
  0.180901 seconds (342.48 k allocations: 26.159 MiB, 5.75% gc time, 15.56% compilation time: 35% of which was recompilation)
```

this PR on julia nightly:
```
julia> @time @eval using QuantumOpticsBase
  0.278220 seconds (563.56 k allocations: 37.528 MiB, 3.67% gc time, 11.03% compilation time: 31% of which was recompilation)
```

Yeah, time to load the library indeed doubles, now to 0.5 seconds, but then julia 1.10 and 1.11 bring it back down. This is regrettable,  but worthwhile for all the stability and performance gained from it.

I will merge this shortly.